@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @package    Grav.Core
  *
@@ -7,6 +8,7 @@
  */
 
 namespace Grav;
+
 define('GRAV_PHP_MIN', '5.6.4');
 
 // Ensure vendor libraries exist
@@ -29,7 +31,7 @@ if (version_compare($ver = PHP_VERSION, $req = GRAV_PHP_MIN, '<')) {
 }
 
 // Register the auto-loader.
-$loader = require_once $autoload;
+$loader = require $autoload;
 
 // Set timezone to default, falls back to system if php.ini not set
 date_default_timezone_set(@date_default_timezone_get());
