@@ -15,10 +15,12 @@ class Session extends \Grav\Framework\Session\Session
 
     /**
      * @return \Grav\Framework\Session\Session
-     * @deprecated 1.5
+     * @deprecated 1.5 Use getInstance() method instead
      */
     public static function instance()
     {
+        user_error(__CLASS__ . '::' . __FUNCTION__ . '() is deprecated since Grav 1.5, use getInstance() method instead', E_USER_DEPRECATED);
+
         return static::getInstance();
     }
 
@@ -51,10 +53,12 @@ class Session extends \Grav\Framework\Session\Session
      * Returns attributes.
      *
      * @return array Attributes
-     * @deprecated 1.5
+     * @deprecated 1.5 Use getAll() method instead
      */
     public function all()
     {
+        user_error(__CLASS__ . '::' . __FUNCTION__ . '() is deprecated since Grav 1.5, use getAll() method instead', E_USER_DEPRECATED);
+
         return $this->getAll();
     }
 
@@ -62,10 +66,12 @@ class Session extends \Grav\Framework\Session\Session
      * Checks if the session was started.
      *
      * @return Boolean
-     * @deprecated 1.5
+     * @deprecated 1.5 Use isStarted() method instead
      */
     public function started()
     {
+        user_error(__CLASS__ . '::' . __FUNCTION__ . '() is deprecated since Grav 1.5, use isStarted() method instead', E_USER_DEPRECATED);
+
         return $this->isStarted();
     }
 
