@@ -14,6 +14,8 @@ namespace Grav\Framework\Session;
  */
 class Session implements SessionInterface
 {
+    protected $options;
+
     /**
      * @var bool
      */
@@ -182,7 +184,10 @@ class Session implements SessionInterface
             unset($_COOKIE[session_name()]);
         }
 
-        $options = $readonly ? ['read_and_close' => '1'] : [];
+        $options = $this->options;
+        if ($readonly) {
+            $options['read_and_close'] = '1';
+        }
 
         $success = @session_start($options);
         if (!$success) {
@@ -335,6 +340,7 @@ class Session implements SessionInterface
             $value = (string)$value;
         }
 
+        $this->options[$key] = $value;
         ini_set($key, $value);
     }
 }
